# apps/server/internal/api/swaggerdocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Robot Center 관제 서버 API입니다. Robot, Operator, Recorder, System 역할별 API를 제공합니다.",
        "title": "Robot Center 관제 API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/operator/recordings": {
            "get": {
                "description": "관제 UI가 조회하는 recording chunk와 파일 상태 목록을 반환합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator API"
                ],
                "summary": "녹화 chunk 조회",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.OperatorRecordingsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/recorder/chunks/{chunkID}/files/{fileType}/uploaded": {
            "post": {
                "description": "recorder-worker가 chunk의 개별 파일 업로드 완료를 보고합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recorder API"
                ],
                "summary": "녹화 파일 업로드 완료",
                "parameters": [
                    {
                        "type": "string",
                        "description": "recording chunk ID",
                        "name": "chunkID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "업로드된 파일 타입",
                        "name": "fileType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "업로드 메타데이터",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RecorderUploadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RecorderRecordingChunkEnvelopeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/recorder/chunks/{chunkID}/uploaded": {
            "post": {
                "description": "recorder-worker가 chunk manifest 업로드 완료를 보고합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recorder API"
                ],
                "summary": "녹화 chunk 업로드 완료",
                "parameters": [
                    {
                        "type": "string",
                        "description": "recording chunk ID",
                        "name": "chunkID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "업로드 메타데이터",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RecorderUploadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RecorderRecordingChunkEnvelopeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/recorder/finalization-jobs/claim": {
            "post": {
                "description": "recorder-worker가 처리할 finalization job을 claim합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recorder API"
                ],
                "summary": "녹화 finalization job claim",
                "parameters": [
                    {
                        "description": "finalization job claim 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RecorderFinalizationClaimRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RecorderFinalizationJobsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/recorder/finalization-jobs/{jobID}/completed": {
            "post": {
                "description": "recorder-worker가 finalization job 완료를 보고합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recorder API"
                ],
                "summary": "녹화 finalization job 완료",
                "parameters": [
                    {
                        "type": "string",
                        "description": "finalization job ID",
                        "name": "jobID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "finalization job 상태 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RecorderFinalizationStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.OKResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/recorder/finalization-jobs/{jobID}/failed": {
            "post": {
                "description": "recorder-worker가 finalization job 실패를 보고합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recorder API"
                ],
                "summary": "녹화 finalization job 실패",
                "parameters": [
                    {
                        "type": "string",
                        "description": "finalization job ID",
                        "name": "jobID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "finalization job 상태 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RecorderFinalizationStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.OKResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/recorder/finalization-jobs/{jobID}/partial": {
            "post": {
                "description": "recorder-worker가 finalization job 부분 완료를 보고합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recorder API"
                ],
                "summary": "녹화 finalization job 부분 완료",
                "parameters": [
                    {
                        "type": "string",
                        "description": "finalization job ID",
                        "name": "jobID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "finalization job 상태 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RecorderFinalizationStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.OKResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/recorder/recording-targets": {
            "get": {
                "description": "recorder-worker가 구독해야 하는 active mission 목록을 반환합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recorder API"
                ],
                "summary": "녹화 대상 임무 조회",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RecorderRecordingTargetsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/recorder/tick": {
            "post": {
                "description": "recorder-worker가 mission/robot 기준 chunk 생성을 요청합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recorder API"
                ],
                "summary": "녹화 tick 반영",
                "parameters": [
                    {
                        "description": "녹화 tick 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RecorderTickRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RecorderRecordingTickResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "robot-center_apps_server_internal_api_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "robot-center_apps_server_internal_api_dto.OKResponse": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                }
            }
        },
...
        "robot-center_apps_server_internal_api_dto.OperatorRecordingChunkResponse": {
            "type": "object",
            "properties": {
                "availableFileTypes": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "boolean"
                    }
                },
                "chunkIndex": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "durationSeconds": {
                    "type": "integer"
                },
                "endedAt": {
                    "type": "string"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.OperatorRecordingFileResponse"
                    }
                },
                "id": {
                    "type": "string"
                },
                "manifestObjectKey": {
                    "type": "string"
                },
                "mediaObjectKeys": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "missionCode": {
                    "type": "string"
                },
                "missionId": {
                    "type": "string"
                },
                "recordingSessionId": {
                    "type": "string"
                },
                "robotCode": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "robot-center_apps_server_internal_api_dto.OperatorRecordingsResponse": {
            "type": "object",
            "properties": {
                "recordings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.OperatorRecordingChunkResponse"
                    }
                }
            }
        },
...
        "robot-center_apps_server_internal_api_dto.RecorderFinalizationClaimRequest": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "lockDurationSeconds": {
                    "type": "integer"
                },
                "workerId": {
                    "type": "string"
                }
            }
        },
...
        "robot-center_apps_server_internal_api_dto.RecorderFinalizationJobsResponse": {
            "type": "object",
            "properties": {
                "jobs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RecorderFinalizationJobResponse"
                    }
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.RecorderFinalizationStatusRequest": {
            "type": "object",
            "properties": {
                "attempt": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "workerId": {
                    "type": "string"
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.RecorderRecordingChunkEnvelopeResponse": {
            "type": "object",
            "properties": {
                "chunk": {
                    "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RecorderRecordingChunkResponse"
                }
            }
        },
...
        "robot-center_apps_server_internal_api_dto.RecorderRecordingTargetsResponse": {
            "type": "object",
            "properties": {
                "targets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RecorderRecordingTargetResponse"
                    }
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.RecorderRecordingTickResponse": {
            "type": "object",
            "properties": {
                "chunk": {
                    "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RecorderRecordingChunkResponse"
                },
                "manifest": {
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.RecorderTickRequest": {
            "type": "object",
            "properties": {
                "chunkDurationSeconds": {
                    "type": "integer"
                },
                "missionCode": {
                    "type": "string"
                },
                "robotCode": {
                    "type": "string"
                },
                "tickAt": {
                    "type": "string"
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.RecorderUploadRequest": {
            "type": "object",
            "properties": {
                "attempt": {
                    "type": "integer"
                },
                "checksum": {
                    "type": "string"
                },
                "sizeBytes": {
                    "type": "integer"
                },
                "workerId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "RobotBearerAuth": {
            "description": "Robot API에서 사용하는 Bearer robotToken입니다.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}